+++ src/components/settings/providers/provider-openai-like.tsx
@@ -1,9 +1,9 @@
 import { observer } from "mobx-react-lite";
 import React from "react";
-import { OPENAI_API_URL } from "../../models/openai";
-import { TTSPluginSettingsStore } from "../../player/TTSPluginSettings";
-import { ApiKeyComponent } from "./api-key-component";
-import { TextInputSetting } from "./setting-components";
+import { OPENAI_API_URL } from "../../../models/openai";
+import { TTSPluginSettingsStore } from "../../../player/TTSPluginSettings";
+import { ApiKeyComponent } from "../api-key-component";
+import { TextInputSetting } from "../setting-components";
 
 export const OpenAICompatibleSettings = observer(
   ({ store }: { store: TTSPluginSettingsStore }) => {
